+++ src/kernel/sections.c
@@ -12,9 +12,9 @@
 #define SECTION_IMPL(section)                        \
     extern const char CONCAT_(__##section, START)[]; \
     extern const char CONCAT_(__##section, END)[];   \
-    const uint64_t CONCAT_(section, START) =         \
-        (uint64_t)CONCAT_(__##section, START);       \
-    const uint64_t CONCAT_(section, END) = (uint64_t)CONCAT_(__##section, END)
+    const uintptr_t CONCAT_(section, START) =        \
+        (uintptr_t)CONCAT_(__##section, START);      \
+    const uintptr_t CONCAT_(section, END) = (uintptr_t)CONCAT_(__##section, END)
 
 SECTION_IMPL(MEMORY);
 SECTION_IMPL(TEXT);
@@ -27,17 +27,17 @@ SECTION_IMPL(USER_user);
 SECTION_IMPL(USER_user2);
 
 extern const char __GLOBAL_SPECIAL_PAGE[];
-const uint64_t GLOBAL_SPECIAL_PAGE = (uint64_t)__GLOBAL_SPECIAL_PAGE;
+const uintptr_t GLOBAL_SPECIAL_PAGE = (uintptr_t)__GLOBAL_SPECIAL_PAGE;
 extern const char __USER_SPECIAL_PAGE[];
-const uint64_t USER_SPECIAL_PAGE = (uint64_t)__USER_SPECIAL_PAGE;
+const uintptr_t USER_SPECIAL_PAGE = (uintptr_t)__USER_SPECIAL_PAGE;
 
 #define PRINT_SECTION(section)                                              \
     print_section(#section, CONCAT_(section, START), CONCAT_(section, END), \
                   SECTION_SIZE(section))
 
-void print_section(char* name, uint64_t start, uint64_t end, uint64_t size) {
-    uint64_t kb = INT_DIV_CEIL(size, 1024);
-    uint64_t pages = INT_DIV_CEIL(size, PAGE_SIZE);
+void print_section(char* name, uintptr_t start, uintptr_t end, uintptr_t size) {
+    uintptr_t kb = INT_DIV_CEIL(size, 1024);
+    uintptr_t pages = INT_DIV_CEIL(size, PAGE_SIZE);
     printf("section %s\n", name);
     printf("\taddresses:  %p - %p\n", start, end);
     printf("\tsize:       %lu pages (~%lu kB, %lu bytes)\n", pages, kb, size);
