+++ src/libc/stddef.h
@@ -4,6 +4,6 @@
 
 #define NULL ((void*)0)
 
-typedef uint64_t size_t;
+typedef uintptr_t size_t;
 
 #define offsetof(type, member) __builtin_offsetof(type, member)
